+++ src/reducers/rootReducer.js
@@ -5,6 +5,7 @@ import auth from './auth';
 import statementBrowser from './statementBrowser';
 import pdfTextAnnotation from './pdfTextAnnotation';
 import viewPaper from './viewPaper';
+import pdfAnnotation from './pdfAnnotation';
 import addTemplate from './addTemplate';
 
 export default history =>
@@ -13,6 +14,7 @@ export default history =>
         addPaper,
         viewPaper,
         statementBrowser,
+        pdfAnnotation,
         auth,
         addTemplate,
         pdfTextAnnotation