+++ src/routes.config.js
@@ -37,6 +37,7 @@ import UserSettings from 'pages/UserSettings';
 import UserProfile from 'pages/UserProfile';
 import FeaturedComparisons from 'pages/FeaturedComparisons';
 import PdfTextAnnotation from 'pages/PdfTextAnnotation';
+import PdfAnnotation from 'pages/PdfAnnotation';
 import { reverse } from 'named-urls';
 
 const routes = [
@@ -223,6 +224,10 @@ const routes = [
     {
         path: ROUTES.TPDL,
         component: () => <Redirect to="/" />
+    },
+    {
+        path: ROUTES.PDF_ANNOTATION,
+        component: PdfAnnotation
     },
     /* Don't add routes below this line */
     {