+++ CHANGELOG.md
@@ -3,6 +3,37 @@
 All notable changes to the ORKG will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and we adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## [0.17](https://gitlab.com/TIBHannover/orkg/orkg-frontend/compare/V0.16...V0.17) - 2020-08-18
+
+### Merged
+
+- Autocomplete improvements [`#346`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/merge_requests/346)
+- fix:  Prevent Scrolling on contribution selection change (ViewPaper) [`#341`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/merge_requests/341)
+- fix: editing issue with Edge [`#347`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/merge_requests/347)
+- fix: shifting page when modal opens/closes [`#345`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/merge_requests/345)
+- feat: PDF survey table extractor [`#314`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/merge_requests/314)
+- fix: Hashtag Search does not work in search page [`#338`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/merge_requests/338)
+- ux: change label of 'published in' [`#335`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/merge_requests/335)
+- fix: empty venue in add paper wizard [`#334`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/merge_requests/334)
+- style(ResearchProblem): show contribution label in papers list [`#340`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/merge_requests/340)
+- fix(GraphView): never automatically expand research fields [`#336`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/merge_requests/336)
+- refactor: move fixed IDs out of .env file to separate config file [`#332`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/merge_requests/332)
+
+### Fixed
+
+- Merge branch 'Autocomplete' into 'master' [`#201`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/issues/201)
+- Merge branch '289-prevent-scrolling-on-contribution-selection-change-viewpaper' into 'master' [`#289`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/issues/289)
+- Merge branch '293-problem-with-editing-papers-while-opening-orkg-web-page-in-different-browsers' into 'master' [`#293`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/issues/293)
+- Merge branch '292-page-moving-when-modal-is-opened' into 'master' [`#292`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/issues/292)
+- Merge branch 'pdf-annotation' into 'master' [`#252`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/issues/252)
+- Merge branch '270-hashtag-search-does-not-work-in-search-page' into 'master' [`#270`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/issues/270)
+- Merge branch '285-replace-label-paper-venue' into 'master' [`#285`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/issues/285)
+- Merge branch '283-add-paper-handeling-empty-venues' into 'master' [`#283`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/issues/283)
+- Merge branch '286-list-contribution-label-in-paper-list' into 'master' [`#286`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/issues/286)
+- Merge branch '284-graph-view-never-automatically-expand-research-fields' into 'master' [`#284`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/issues/284)
+- Merge branch '281-move-graph-ids-to-config-files' into 'master' [`#281`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/issues/281)
+
+---
 ## [0.16](https://gitlab.com/TIBHannover/orkg/orkg-frontend/compare/V0.15...V0.16) - 2020-07-22
 
 ### Merged
@@ -211,7 +242,7 @@ All notable changes to the ORKG will be documented in this file. The format is b
 - Fixes #185. Published comparison url [`#185`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/issues/185)
 
 ---
-## [0.9](https://gitlab.com/TIBHannover/orkg/orkg-frontend/compare/V0.8...V0.9) - 2020-05-05
+## [0.9](https://gitlab.com/TIBHannover/orkg/orkg-frontend/compare/V0.8...V0.9) - 2020-04-24
 
 ### Merged
 
@@ -296,6 +327,8 @@ All notable changes to the ORKG will be documented in this file. The format is b
 ### Commits
 
 - Fix datacube vocabulary tabular view [`1e770d8`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/commit/1e770d8a25143736794893550ea3d869c1efed5a)
+- Include public widget file [`3f119d8`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/commit/3f119d86126a764d8951ab648c0f1af0bdbe6f0a)
+- Fix problem with fetching classes [`fa6223c`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/commit/fa6223cd715f7b30283c46459705d184aab6c8a7)
 ---
 ## [0.6](https://gitlab.com/TIBHannover/orkg/orkg-frontend/compare/V0.5.8...V0.6) - 2020-03-10
 
@@ -386,6 +419,7 @@ All notable changes to the ORKG will be documented in this file. The format is b
 ### Commits
 
 - fix(Comparison): Export as RDF [`c6abbd6`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/commit/c6abbd667fac2bcb441360dac68925ff12bd1674)
+- Update changelog [`fc46dbd`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/commit/fc46dbd30c79d7a0ad6af8ef5c8e0ca9b658c61c)
 - fix(Comparison): Comparison url [`27f3e7b`](https://gitlab.com/TIBHannover/orkg/orkg-frontend/commit/27f3e7b7966de5fb28bdb197919a3fd5e49015de)
 ---
 ## [0.5.4](https://gitlab.com/TIBHannover/orkg/orkg-frontend/compare/V0.5.3...V0.5.4) - 2020-01-17